js files modified

+++ controllers/testimony-controller.js
@@ -24,27 +24,74 @@ export const postTestimony = async (req, res, next) => {
     }
 }
 
+// Approve testimony for only admins
+export const approveTestimony = async (req, res, next) => {
+    try {
+        // Check if the user is an admin
+        if (req.auth.role !== "admin") {
+            return res.status(403).json("You do not have permission to approve this testimony!");
+        }
+        const testimony = await TestimonyModel.findOneAndUpdate(req.params.id);
+        if (!testimony) {
+            return res.status(404).json("Testimony not found!");
+        }
+        if (testimony.approved) {
+            return res.status(400).json("Testimony is already approved!");
+        }
+        testimony.approved = true;
+        await testimony.save();
+        // Respond to request
+        res.status(200).json("Testimony approved!");
+    } catch (error) {
+        next(error);
+    }
+}
+
 // Get all testimonies
 export const getAllTestimonies = async (req, res, next) => {
     try {
         const { filter = "{}", sort = "{}", limit = 50, skip = 0 } = req.query;
-        // Fetch reservations from database 
+        let query = { ...JSON.parse(filter) };
+        // Check if the user is an admin
+        if (req.auth.role !== "admin") {
+            query.approved = true;
+        }
+        // Fetch testimonies from database 
         const testimonies = await TestimonyModel
-            .find(filter)
+            .find(query)
             .sort(JSON.parse(sort))
             .limit(limit)
             .skip(skip);
-        // Return a response
+        // Respond to request
         res.status(200).json(testimonies);
     } catch (error) {
         next(error);
     }
 }
 
-// delete an existing testimony
+// Get a testimony by id
+export const getTestimonyById = async (req, res, next) => {
+    try {
+        // Fetch a testimony from the database
+        const testimony = await TestimonyModel.findById(req.params.id);
+        if (!testimony) {
+            return res.status(404).json("Testimony not found!");
+        }
+        // Return Response
+        res.status(200).json(testimony);
+    } catch (error) {
+        next(error);
+    }
+}
+
+// Delete an existing testimony
 export const deleteTestimony = async (req, res, next) => {
     try {
-        // Find a reservation by id and delete it
+        // Check if the user is an admin
+        if (req.auth.role !== "admin") {
+            return res.status(403).json("You do not have permission to delete this testimony!");
+        }
+        // Find a testimony by id and delete it
         const testimony = await TestimonyModel.
             findOneAndDelete(
                 {
@@ -55,6 +102,7 @@ export const deleteTestimony = async (req, res, next) => {
         if (!testimony) {
             return res.status(404).json("Testimony not found!");
         }
+        // Respond to request
         res.status(200).json("Testimony deleted!");
     } catch (error) {
         next(error);
